--- a/metrics.go
+++ b/metrics.go
@@ -8,9 +8,9 @@ import (
 )
 
 type Metrics interface {
-	WriteTxnTableAcquisition(tableName string, acquire time.Duration)
-	WriteTxnTotalAcquisition(goPackage string, tables []string, acquire time.Duration)
-	WriteTxnDuration(goPackage string, acquire time.Duration)
+	WriteTxnTableAcquisition(handle string, tableName string, acquire time.Duration)
+	WriteTxnTotalAcquisition(handle string, tables []string, acquire time.Duration)
+	WriteTxnDuration(handle string, tables []string, acquire time.Duration)
 
 	GraveyardLowWatermark(tableName string, lowWatermark Revision)
 	GraveyardCleaningDuration(tableName string, duration time.Duration)
@@ -121,16 +121,16 @@ func (m *ExpVarMetrics) ObjectCount(name string, numObjects int) {
 	m.ObjectCountVar.Set(name, &intVar)
 }
 
-func (m *ExpVarMetrics) WriteTxnDuration(goPackage string, acquire time.Duration) {
-	m.WriteTxnDurationVar.AddFloat(goPackage, acquire.Seconds())
+func (m *ExpVarMetrics) WriteTxnDuration(handle string, tables []string, acquire time.Duration) {
+	m.WriteTxnDurationVar.AddFloat(handle+"/"+strings.Join(tables, "+"), acquire.Seconds())
 }
 
-func (m *ExpVarMetrics) WriteTxnTotalAcquisition(goPackage string, tables []string, acquire time.Duration) {
-	m.WriteTxnAcquisitionVar.AddFloat(goPackage, acquire.Seconds())
+func (m *ExpVarMetrics) WriteTxnTotalAcquisition(handle string, tables []string, acquire time.Duration) {
+	m.WriteTxnAcquisitionVar.AddFloat(handle+"/"+strings.Join(tables, "+"), acquire.Seconds())
 }
 
-func (m *ExpVarMetrics) WriteTxnTableAcquisition(name string, acquire time.Duration) {
-	m.LockContentionVar.AddFloat(name, acquire.Seconds())
+func (m *ExpVarMetrics) WriteTxnTableAcquisition(handle string, tableName string, acquire time.Duration) {
+	m.LockContentionVar.AddFloat(handle+"/"+tableName, acquire.Seconds())
 }
 
 var _ Metrics = &ExpVarMetrics{}
@@ -162,15 +162,15 @@ func (*NopMetrics) Revision(tableName string, revision uint64) {
 }
 
 // WriteTxnDuration implements Metrics.
-func (*NopMetrics) WriteTxnDuration(goPackage string, acquire time.Duration) {
+func (*NopMetrics) WriteTxnDuration(handle string, tables []string, acquire time.Duration) {
 }
 
 // WriteTxnTableAcquisition implements Metrics.
-func (*NopMetrics) WriteTxnTableAcquisition(tableName string, acquire time.Duration) {
+func (*NopMetrics) WriteTxnTableAcquisition(handle string, tableName string, acquire time.Duration) {
 }
 
 // WriteTxnTotalAcquisition implements Metrics.
-func (*NopMetrics) WriteTxnTotalAcquisition(goPackage string, tables []string, acquire time.Duration) {
+func (*NopMetrics) WriteTxnTotalAcquisition(handle string, tables []string, acquire time.Duration) {
 }
 
 var _ Metrics = &NopMetrics{}